# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups": {
      "post": {
        "operationId": "GroupsController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an Equb group",
        "tags": [
          "Groups"
        ]
      },
      "get": {
        "operationId": "GroupsController_listGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupSummaryResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List groups where current user is an active member",
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/join": {
      "post": {
        "operationId": "GroupsController_joinGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupJoinResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invite is invalid or unusable"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Removed members cannot self-rejoin"
          },
          "404": {
            "description": "Invite code not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join a group with invite code",
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{id}": {
      "get": {
        "operationId": "GroupsController_getGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active group membership required"
          },
          "404": {
            "description": "Group or membership not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get group details for current member",
        "tags": [
          "Groups",
          "Members"
        ]
      }
    },
    "/groups/{id}/invite": {
      "post": {
        "operationId": "GroupsController_createInvite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid invite constraints"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active admin membership required"
          },
          "404": {
            "description": "Group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create invite code for group",
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{id}/members": {
      "get": {
        "operationId": "GroupsController_listMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupMemberResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active group membership required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List members in a group",
        "tags": [
          "Groups",
          "Members"
        ]
      }
    },
    "/groups/{id}/members/{userId}/role": {
      "patch": {
        "operationId": "GroupsController_updateMemberRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupMemberResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Last active admin cannot be removed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active admin membership required"
          },
          "404": {
            "description": "Member not found in group"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update member role in a group",
        "tags": [
          "Groups",
          "Members"
        ]
      }
    },
    "/groups/{id}/members/{userId}/status": {
      "patch": {
        "operationId": "GroupsController_updateMemberStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupMemberResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Status transition rule violation"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Only self-leave or admin remove is allowed"
          },
          "404": {
            "description": "Member not found in group"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update member status (self leave or admin remove)",
        "tags": [
          "Groups",
          "Members"
        ]
      }
    },
    "/groups/{id}/payout-order": {
      "patch": {
        "operationId": "GroupsController_updatePayoutOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PayoutOrderItemDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupMemberResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Payout positions must be contiguous and unique"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active admin membership required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set payout order for active members",
        "tags": [
          "Groups",
          "Cycles"
        ]
      }
    },
    "/groups/{id}/cycles/generate": {
      "post": {
        "operationId": "GroupsController_generateCycles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCyclesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupCycleResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cycle generation constraints not satisfied"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active admin membership required"
          },
          "404": {
            "description": "Group not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate next cycle(s) based on payout order",
        "tags": [
          "Groups",
          "Cycles"
        ]
      }
    },
    "/groups/{id}/cycles/current": {
      "get": {
        "operationId": "GroupsController_getCurrentCycle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupCycleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active group membership required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current open cycle for a group",
        "tags": [
          "Groups",
          "Cycles"
        ]
      }
    },
    "/groups/{id}/cycles/{cycleId}": {
      "get": {
        "operationId": "GroupsController_getCycleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupCycleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active group membership required"
          },
          "404": {
            "description": "Cycle not found in group"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get cycle details by id",
        "tags": [
          "Groups",
          "Cycles"
        ]
      }
    },
    "/groups/{id}/cycles": {
      "get": {
        "operationId": "GroupsController_listCycles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupCycleResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Active group membership required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List cycles for a group",
        "tags": [
          "Groups",
          "Cycles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Equb API",
    "description": "Equb backend API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Family Equb"
          },
          "contributionAmount": {
            "type": "number",
            "example": 500
          },
          "frequency": {
            "type": "string",
            "enum": [
              "WEEKLY",
              "MONTHLY"
            ],
            "example": "MONTHLY"
          },
          "startDate": {
            "type": "string",
            "example": "2026-03-01"
          },
          "currency": {
            "type": "string",
            "example": "ETB"
          }
        },
        "required": [
          "name",
          "contributionAmount",
          "frequency",
          "startDate"
        ]
      },
      "CurrentMembershipResponseDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "INVITED",
              "ACTIVE",
              "LEFT",
              "REMOVED"
            ]
          }
        },
        "required": [
          "role",
          "status"
        ]
      },
      "GroupDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "contributionAmount": {
            "type": "number"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "WEEKLY",
              "MONTHLY"
            ]
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ARCHIVED"
            ]
          },
          "createdByUserId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "strictPayout": {
            "type": "boolean"
          },
          "timezone": {
            "type": "string"
          },
          "membership": {
            "$ref": "#/components/schemas/CurrentMembershipResponseDto"
          }
        },
        "required": [
          "id",
          "name",
          "currency",
          "contributionAmount",
          "frequency",
          "startDate",
          "status",
          "createdByUserId",
          "createdAt",
          "strictPayout",
          "timezone",
          "membership"
        ]
      },
      "GroupSummaryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "contributionAmount": {
            "type": "number"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "WEEKLY",
              "MONTHLY"
            ]
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ARCHIVED"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "currency",
          "contributionAmount",
          "frequency",
          "startDate",
          "status"
        ]
      },
      "JoinGroupDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "A1B2C3D4"
          }
        },
        "required": [
          "code"
        ]
      },
      "GroupJoinResponseDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "INVITED",
              "ACTIVE",
              "LEFT",
              "REMOVED"
            ]
          },
          "joinedAt": {
            "type": "object"
          }
        },
        "required": [
          "groupId",
          "role",
          "status"
        ]
      },
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "string",
            "example": "2026-04-01T00:00:00.000Z"
          },
          "maxUses": {
            "type": "number",
            "example": 20
          }
        }
      },
      "InviteCodeResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "joinUrl": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "joinUrl"
        ]
      },
      "MemberProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "fullName": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "phone"
        ]
      },
      "GroupMemberResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/MemberProfileResponseDto"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "INVITED",
              "ACTIVE",
              "LEFT",
              "REMOVED"
            ]
          },
          "payoutPosition": {
            "type": "object",
            "nullable": true
          },
          "joinedAt": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "user",
          "role",
          "status"
        ]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "example": "ADMIN"
          }
        },
        "required": [
          "role"
        ]
      },
      "UpdateMemberStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "LEFT",
              "REMOVED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "PayoutOrderItemDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "payoutPosition": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "userId",
          "payoutPosition"
        ]
      },
      "GenerateCyclesDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "default": 1,
            "minimum": 1,
            "maximum": 12
          }
        }
      },
      "CyclePayoutUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "fullName": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "phone"
        ]
      },
      "GroupCycleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "cycleNo": {
            "type": "number"
          },
          "dueDate": {
            "format": "date-time",
            "type": "string"
          },
          "payoutUserId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "CLOSED"
            ]
          },
          "createdByUserId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "payoutUser": {
            "$ref": "#/components/schemas/CyclePayoutUserResponseDto"
          }
        },
        "required": [
          "id",
          "groupId",
          "cycleNo",
          "dueDate",
          "payoutUserId",
          "status",
          "createdByUserId",
          "createdAt",
          "payoutUser"
        ]
      },
...
    }
  }
}